--- testing/main.cpp
@@ -1,8 +1,0 @@
-#include "../testing/testHeader.h"
-
-int main(void){
-    testRute();
-    testCity();
-    srand(1);
-    testPopulation(new Population(300));
-}
